feat: redesign chat style and functionality #158
