+++ frontend/app/_components/CamyoNavBar.jsx
@@ -93,8 +93,8 @@ export default function CamyoWebNavBar() {
             </View>
             <View style={styles.rightSection}>
                
-               <TouchableOpacity onPress={() => router.replace('/chat/list')}>
-                  <Text style={styles.linkText}>Go to Chat List</Text>
+               <TouchableOpacity onPress={() => router.replace('/chat')}>
+                  <Text style={styles.linkText}>Chats</Text>
                </TouchableOpacity>
                <TouchableOpacity style={styles.buttonText} onPress={() => router.push(routes.listcompanies)} ><Text style={styles.linkText}>Empresas</Text></TouchableOpacity>
