+++ frontend/app/chat/index.tsx
@@ -1,189 +1,173 @@
-import React, { useState, useEffect, useCallback } from 'react';
-import { GiftedChat, InputToolbar, MessageText } from 'react-native-gifted-chat';
-import { 
-    collection, 
-    addDoc, 
-    query, 
-    where, 
-    getDocs, 
-    orderBy, 
-    onSnapshot, 
-    updateDoc, 
-    doc, 
-    getDoc 
-} from 'firebase/firestore';
-import { database } from '../../firebase';
+import React, { useState, useEffect } from 'react';
+import { View, Text, FlatList, TouchableOpacity, StyleSheet } from 'react-native';
 import { useAuth } from '../../contexts/AuthContext';
-import { useSearchParams } from 'expo-router/build/hooks';
+import ChatComponent from './chat';  // Componente de chat
+import { database } from '../../firebase';
+import { collection, query, where, getDocs, onSnapshot } from 'firebase/firestore';
+import colors from '@/assets/styles/colors';
 import { useRouter } from 'expo-router';
-import withNavigationGuard from '@/hoc/withNavigationGuard';
-import ProtectedRoute from '../../security/ProtectedRoute';
+import ProtectedRoute from '@/security/ProtectedRoute';
 
-interface Message {
-    _id: string;
-    createdAt: Date;
-    text: string;
-    seen: boolean;
-    user: {
-        _id: string;
-        name?: string;
-        avatar?: string;
-    };
+interface Chat {
+  id: string;
+  participants: string[];
+  lastMessage: string;
+  unreadMessagesCount: number;
 }
 
-function Chat({ route }: any) {
-    const { user } = useAuth();
-    const router = useRouter();
-    const searchParams = useSearchParams();
-    const otherUserId = searchParams.get('otherUserId')?.toString();
+interface Usuario {
+  userId: string;
+  nombre: string;
+  foto: string;
+}
 
-    const [messages, setMessages] = useState<Message[]>([]);
-    const [chatId, setChatId] = useState<string | null>(null);
+function ChatList() {
+  const { user } = useAuth();
+  const [chats, setChats] = useState<Chat[]>([]);
+  const [userDetails, setUserDetails] = useState<{ [key: string]: Usuario }>({});
+  const [currentChat, setCurrentChat] = useState<Chat | null>(null);  // Estado para el chat seleccionado
+  const BACKEND_URL = process.env.EXPO_PUBLIC_BACKEND_URL;
 
-    const sendMessage = async (chatId: string, messageText: string) => {
-        try {
-            const chatDocRef = doc(database, 'chats', chatId);
-            await updateDoc(chatDocRef, {
-                lastMessage: messageText, 
-                lastUpdated: new Date(), 
-            });
-        } catch (error) {
-            console.error('Error updating chat:', error);
-        }
-    };
-
-    const getOrCreateChat = async (userId1: string, userId2: string) => {
-        const chatsRef = collection(database, 'chats');
-        const q = query(chatsRef, where('participants', 'array-contains', userId1));
-        const querySnapshot = await getDocs(q);
-
-        let chat = null;
-
-        querySnapshot.forEach(doc => {
-            const data = doc.data();
-            if (data.participants.includes(userId2)) {
-                chat = { id: doc.id, ...data };
-            }
+  useEffect(() => {
+    if (!user || !user?.userId) return;
+  
+    const chatsRef = collection(database, 'chats');
+    const q = query(chatsRef, where('participants', 'array-contains', user.userId.toString()));
+  
+    const unsubscribe = onSnapshot(q, (querySnapshot) => {
+      if (querySnapshot.empty) {
+        setChats([]); // Asegurar que se vacíe la lista si no hay chats
+        return;
+      }
+  
+      const fetchedChats: Chat[] = [];
+  
+      querySnapshot.forEach((doc) => {
+        const data = doc.data();
+  
+        fetchedChats.push({
+          id: doc.id,
+          participants: data.participants,
+          lastMessage: data.lastMessage?.trim() !== "" ? data.lastMessage : "Inicia una conversación",
+          unreadMessagesCount: data.unreadMessagesCount || 0,
+          lastUpdated: data.lastUpdated || new Date(0), // Asegúrate de que el campo existe y es una fecha
         });
+      });
+  
+      // Ordenar los chats por la fecha de `lastUpdated` de manera descendente (más reciente primero)
+      fetchedChats.sort((a, b) => b.lastUpdated - a.lastUpdated);
+  
+      setChats(fetchedChats);
+    });
+  
+    return () => unsubscribe(); // Se desuscribe cuando el componente se desmonta
+  }, [user?.userId]);
+  
+  
+  
+  const handleChatClick = (chat: Chat) => {
+    setCurrentChat(chat);  // Establecer el chat seleccionado en el estado
+  };
 
-        if (chat) {
-            return chat;
-        }
-
-        const newChat = await addDoc(chatsRef, {
-            participants: [userId1, userId2],
-        });
-
-        return { id: newChat.id, participants: [userId1, userId2] };
+  useEffect(() => {
+    if (!user?.userId) return;
+    const fetchUserDetails = async (otherUserId: string) => {
+      try {
+        const response = await fetch(`${BACKEND_URL}/usuarios/${otherUserId}`);
+        if (!response.ok) throw new Error('Error fetching user data');
+        const data: Usuario = await response.json();
+        setUserDetails(prev => ({ ...prev, [otherUserId]: data }));
+      } catch (err) {
+        console.error(err);
+      }
     };
 
-    const loadMessages = (chatId: string) => {
-        const messagesRef = collection(database, `chats/${chatId}/messages`);
-        const q = query(messagesRef, orderBy('createdAt', 'desc'));
+    chats.forEach(chat => {
+      const otherUserId = chat.participants.find(participant => participant !== user.userId.toString());
+      if (otherUserId && !userDetails[otherUserId]) {
+        fetchUserDetails(otherUserId);
+      }
+    });
+  }, [chats, user?.userId, userDetails]);
 
-        const unsubscribe = onSnapshot(q, querySnapshot => {
-            const messages = querySnapshot.docs.map(doc => ({
-                _id: doc.id,
-                ...doc.data(),
-                createdAt: doc.data().createdAt.toDate(),
-            }));
+  const renderChatItem = ({ item }: { item: Chat }) => {
+    if (!user?.userId) return;
+    const otherUserId = item.participants.find(participant => participant !== user.userId.toString());
+    const otherUser = otherUserId ? userDetails[otherUserId] : null;
 
-            setMessages(messages);
-
-            // Marcar los mensajes como vistos cuando son cargados
-            messages.forEach(message => {
-                if (user && message.user._id !== user.userId.toString() && !message.seen) {
-                    markMessageAsSeen(chatId, message._id, user.userId.toString());
-                }
-            });
-        });
-
-        return unsubscribe;
-    };
-
-    const markMessageAsSeen = async (chatId: string, messageId: string, userId: string) => {
-        const messageRef = doc(database, `chats/${chatId}/messages`, messageId);
-        const messageSnapshot = await getDoc(messageRef);
-        const messageData = messageSnapshot.data();
-
-        if (messageData && !messageData.seen && messageData.user._id !== userId) {
-            await updateDoc(messageRef, { seen: true });
-        }
-    };
-
-    useEffect(() => {
-        if (!otherUserId) {
-            router.replace("/chat/list");
-            return;
-        }
-        const setupChat = async () => {
-            if (!user || !user.userId || !otherUserId) return;
-
-            const userId = user.userId.toString();
-
-            const chat = await getOrCreateChat(userId, otherUserId);
-            setChatId(chat.id);
-
-            const unsubscribe = loadMessages(chat.id);
-            return () => unsubscribe();
-        };
-
-        setupChat();
-    }, [user?.userId, otherUserId]);
-
-    const onSend = useCallback(
-        async (newMessages: Message[] = []) => {
-            if (!chatId || !user) return;
-
-            setMessages(previousMessages => GiftedChat.append(previousMessages, newMessages));
-            
-            const { _id, createdAt, text, user: messageUser } = newMessages[0];
-
-            await sendMessage(chatId, text);
-
-            addDoc(collection(database, `chats/${chatId}/messages`), {
-                _id,
-                createdAt,
-                text,
-                seen: false, // Inicialmente, el mensaje no está visto
-                user: {
-                    ...messageUser,
-                    _id: messageUser._id.toString(),
-                },
-            });
-        },
-        [chatId, user]
-    );
+    const isSelected = currentChat?.id === item.id;  // Verificar si este chat es el seleccionado
 
     return (
-        <ProtectedRoute>
-        <GiftedChat
-            messages={messages}
-            showAvatarForEveryMessage={false}
-            showUserAvatar={true}
-            onSend={messages => onSend(messages)}
-            messagesContainerStyle={{
-                backgroundColor: '#fff',
-            }}
-            renderInputToolbar={props => (
-                <InputToolbar
-                    {...props}
-                    containerStyle={{
-                        backgroundColor: '#fff',
-                        borderRadius: 20,
-                    }}
-                />
-            )}
-            user={{
-                _id: user?.userId.toString(),
-                name: user?.nombre,
-                avatar: user?.foto,
-            }}
-            renderMessageText={(props) => <MessageText {...props} />}
-        />
-        </ProtectedRoute>
+      <View style={styles.listContainer}>
+        <TouchableOpacity
+          style={[styles.chatItem, isSelected && styles.selectedChatItem]}  // Aplicar estilo de chat seleccionado
+          onPress={() => handleChatClick(item)}  // Usamos la función para seleccionar el chat
+        >
+          <Text style={styles.chatText}>{otherUser ? otherUser.nombre : `Chat with User ID: ${otherUserId}`}</Text>
+          <Text style={styles.lastMessage}>{item.lastMessage || 'No messages yet'}</Text>
+        </TouchableOpacity>
+      </View>
     );
+  };
+
+  return (
+    <ProtectedRoute>
+    <View style={styles.container}>
+      <FlatList
+        data={chats}
+        keyExtractor={item => item.id}
+        renderItem={renderChatItem}
+        ListEmptyComponent={<Text style={styles.emptyText}>No chats available</Text>}
+      />
+      <View style={{ flex: 10 }}>
+        {currentChat && <ChatComponent chat={currentChat} />}  {/* Mostrar el chat seleccionado */}
+      </View>
+    </View>
+    </ProtectedRoute>
+  );
 }
 
-// Exportamos el componente envuelto con el HOC `withNavigationGuard`
-export default withNavigationGuard(Chat);
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'row',  // Para mostrar los chats y el chat en paralelo
+    paddingTop: 80,
+    backgroundColor: '#fff',
+    marginRight: 0,
+  },
+  listContainer: {
+    width: "100%"
+  },
+  chatItem: {
+    paddingVertical: 0, // Reducir el padding superior e inferior
+    paddingHorizontal: 15,
+    flexBasis: 120,
+    borderBottomWidth: 1,
+    borderBottomColor: '#ccc',
+    width: '100%',
+  },
+  selectedChatItem: {
+    backgroundColor: '#FFE4B5',  // Color de fondo para el chat seleccionado
+    borderLeftWidth: 4,  // Resaltar con borde izquierdo
+    borderLeftColor: colors.secondary,  // Color del borde
+  },
+  unreadChatItem: {
+    backgroundColor: '#f0f8ff',
+  },
+  chatText: {
+    fontSize: 18,
+    fontWeight: 'bold',
+  },
+  lastMessage: {
+    fontSize: 14,
+    color: '#666',
+  },
+  emptyText: {
+    textAlign: 'center',
+    fontSize: 16,
+    color: '#999',
+    marginTop: 20,
+  },
+});
+
+export default ChatList;
